:bug: Handle `analyze: false` alonsgide with a custom format option
If a format option was passed alongside `analyze: false` option it'd generate a wrong string due to a missing `gsub` to handle this case. That would generate a string such as `"(FORMAT JSON, )"` which is invalid SQL.

This fix adds a regex to detect and gsub this case and also a unit test to cover this up.

Closes #3